feat(Files): add SetPos and GetPos procedures to Files module
The Files module has the Seek procedure to allow one to move the file
offset to a certain position. However, it doesn't have any procedure
that gives us the current file offset. Hence, the addition of GetPos.
SetPos is just a wrapper for Seek, and it is more of a copy of
Modula-2's FileSystem.SetPos